linebot_with_django3_1_7/requirements.txt
Django==3.1.7
line-bot-sdk>=2.3.0
gunicorn==20.*
gevent==21.*